Revert "Dockerfile corrections - improved documentation"
