--- payloads_src/MogwaiLabsMJET/src/main/java/de/mogwailabs/MogwaiLabsMJET/MogwaiLabsPayloadMBean.java
@@ -1,10 +1,0 @@
-package de.mogwailabs.MogwaiLabsMJET;
-
-public interface MogwaiLabsPayloadMBean {
-    
-    public abstract String runCMD(String passwd, String cmd, String shell);
-    public abstract String runJS(String passwd, String js);
-    public abstract boolean changePassword(String oldPass, String newPass);
-    
-}
-
