+++ payloads_src/siberas_mlet/src/de/siberas/lab/SiberasPayloadMBean.java
@@ -1,0 +1,9 @@
+ package de.siberas.lab;
+
+public interface SiberasPayloadMBean {
+	
+	public String runCMD(String passwd, String cmd);
+	public String runJS(String passwd, String js);
+	public boolean changePassword(String oldPass, String newPass);
+	
+}
